backend/requirements.txt
openai
python-telegram-bot
requests
schedule
python-dotenv
